Overhaul!!!
> Changed how reasons are built.
  * Replaces any instance of `$ruleName` with the rule.
  * Warn/Mute/Kick/Ban all pass the correct reason, no more empty reasons
> Custom Help Command
  * Embeded results
  * Hidden staff commands, to make easier for members to find available commands
> MessageEvent no longer spam mutes a user.

+++ src/main/java/main/commands/admin/OfftopicCommand.java
@@ -1,21 +1,25 @@
-package main.commands;
+package main.commands.admin;
 
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
 
 import main.lib.Constants;
+import main.lib.Utils;
 
 public class OfftopicCommand extends Command {
 	public OfftopicCommand() {
 		name = "offtopic";
 		help = "redirects members to where offtopic conversation can be had.";
 		guildOnly = true;
-		category = new Category("utility");
-		requiredRole = "staff";
+		category = new Category("staff");
 	}
 	
 	@Override
 	protected void execute(CommandEvent event) {
+		if (!(Utils.hasRoleWithName(event.getMember(), "staff"))) {
+			return;
+		}
+		
 		event.getMessage().delete().queue();
 		event.reply("This is not the place to enjoy off-topic conversation. We advise that you carry your conversation over to " 
 					+ event.getGuild().getTextChannelById(Constants.commonsID)
